refactor: replace DeletePartyDialog with ConfirmDialog
